pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-enterprise-search"
version = "0.1.0"
description = "Secure, multilingual, context-aware enterprise search with RAG"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Enterprise Search Team"}
]
keywords = ["search", "rag", "enterprise", "opensearch", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "opensearch-py>=2.4.0",
    "sentence-transformers>=2.2.2",
    "torch>=2.1.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "tika>=2.6.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.1.0",
    "python-magic>=0.4.27",
    "langdetect>=1.0.9",
    "unstructured[pdf,docx]>=0.11.0",
    "presidio-analyzer>=2.2.33",
    "presidio-anonymizer>=2.2.33",
    "spacy>=3.7.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "prefect>=2.14.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "prometheus-client>=0.19.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
